ORCA-310 Support absolute paths.
This may not be the appropriate fix, but it was the most concise.

ORCA assumes all paths are subpaths to a location in AbstractPathHelper. This has forced ORCA to work with relative paths, which can be a pain. Especially when trying to determine the relative path for a far away directory. This adds support for direct paths.